+++ internal/ui/plugins/loader.go
@@ -4,12 +4,14 @@ import (
 	"github.com/devnullvoid/pvetui/internal/config"
 	"github.com/devnullvoid/pvetui/internal/ui/components"
 	"github.com/devnullvoid/pvetui/internal/ui/plugins/communityscripts"
+	"github.com/devnullvoid/pvetui/internal/ui/plugins/guestlist"
 )
 
 type factory func() components.Plugin
 
 var registry = map[string]factory{
 	communityscripts.PluginID: func() components.Plugin { return communityscripts.New() },
+	guestlist.PluginID:        func() components.Plugin { return guestlist.New() },
 }
 
 var defaultEnabled = []string{}